Move coverage handling to Protein module
