+++ ms2/src/org/labkey/ms2/MS2Manager.java
@@ -76,7 +76,7 @@ import org.labkey.ms2.pipeline.AbstractMS2SearchTask;
 import org.labkey.ms2.pipeline.MS2ImportPipelineJob;
 import org.labkey.ms2.pipeline.TPPTask;
 import org.labkey.ms2.pipeline.mascot.MascotImportPipelineJob;
-import org.labkey.ms2.protein.CoverageProtein;
+import org.labkey.api.protein.CoverageProtein;
 import org.labkey.ms2.protein.Protein;
 import org.labkey.ms2.query.MS2Schema;
 import org.labkey.ms2.reader.ITraqProteinQuantitation;
@@ -1673,13 +1673,13 @@ public class MS2Manager
 
     public static Protein getProtein(int seqId)
     {
-        SimpleProtein simpleProtein = ProteinManager.getProtein(seqId);
+        SimpleProtein simpleProtein = ProteinManager.getSimpleProtein(seqId);
         return simpleProtein != null ? new Protein(simpleProtein) : null;
     }
 
     public static CoverageProtein getCoverageProtein(int seqId)
     {
-        SimpleProtein simpleProtein = ProteinManager.getProtein(seqId);
+        SimpleProtein simpleProtein = ProteinManager.getSimpleProtein(seqId);
         return simpleProtein != null ? new CoverageProtein(simpleProtein) : null;
     }
 }
