+++ ms2/src/org/labkey/ms2/MS2Module.java
@@ -17,7 +17,6 @@ package org.labkey.ms2;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import org.labkey.api.annotations.Migrate;
 import org.labkey.api.data.CompareType;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.ContainerManager;
@@ -33,7 +32,6 @@ import org.labkey.api.module.ModuleContext;
 import org.labkey.api.module.SpringModule;
 import org.labkey.api.ms2.MS2Service;
 import org.labkey.api.pipeline.PipelineService;
-import org.labkey.api.protein.ProteinCoverageViewService;
 import org.labkey.api.protein.ProteinManager;
 import org.labkey.api.protein.ProteinSchema;
 import org.labkey.api.protein.ProteinService;
@@ -41,6 +39,7 @@ import org.labkey.api.protein.ProteomicsModule;
 import org.labkey.api.protein.query.SequencesTableInfo;
 import org.labkey.api.protein.search.PepSearchModel;
 import org.labkey.api.protein.search.ProbabilityProteinSearchForm;
+import org.labkey.api.protein.search.ProteinSearchWebPart;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.query.QueryView;
 import org.labkey.api.reports.ReportService;
@@ -82,7 +81,6 @@ import org.labkey.ms2.pipeline.sequest.SequestSearchTask;
 import org.labkey.ms2.pipeline.sequest.ThermoSequestParamsBuilder;
 import org.labkey.ms2.pipeline.tandem.XTandemPipelineProvider;
 import org.labkey.ms2.protein.Protein;
-import org.labkey.ms2.protein.ProteinCoverageViewServiceImpl;
 import org.labkey.ms2.query.MS2Schema;
 import org.labkey.ms2.reader.DatDocumentParser;
 import org.labkey.ms2.reader.MGFDocumentParser;
@@ -92,7 +90,6 @@ import org.labkey.ms2.reader.PeptideProphetSummary;
 import org.labkey.ms2.reader.RandomAccessJrapMzxmlIterator;
 import org.labkey.ms2.reader.SequestLogDocumentParser;
 import org.labkey.ms2.search.MSSearchWebpart;
-import org.labkey.api.protein.search.ProteinSearchWebPart;
 
 import java.io.File;
 import java.util.Collection;
@@ -181,7 +178,6 @@ public class MS2Module extends SpringModule implements ProteomicsModule
         return true;
     }
 
-    @Migrate
     @Override
     protected void init()
     {
@@ -190,9 +186,6 @@ public class MS2Module extends SpringModule implements ProteomicsModule
 
         MS2Schema.register(this);
         MS2Service.setInstance(new MS2ServiceImpl());
-
-        // TODO: Move to ProteinModule after migrating ProteinCoverageViewServiceImpl
-        ProteinCoverageViewService.setInstance(new ProteinCoverageViewServiceImpl());
     }
 
     @Override
